% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma2eset.R
\name{as.ExpressionSet}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet.WideSomaLogicData}
\alias{as.MSnSet}
\alias{as.MSnSet.WideSomaLogicData}
\alias{soma2eset}
\title{Convert objects into ExpressionSets or MSnSets.}
\usage{
as.ExpressionSet(x, log2Transform = TRUE, ...)

\method{as.ExpressionSet}{WideSomaLogicData}(x, log2Transform = TRUE, ...)

soma2eset(somaObj, log2Transform = TRUE)

\method{as.MSnSet}{WideSomaLogicData}(x, log2Transform = FALSE, ...)

as.MSnSet(x, log2Transform = FALSE, ...)
}
\arguments{
\item{x}{An object to transform.  Currently only \code{WideSomaLogicData}
objects are supported.}

\item{log2Transform}{whether to log2 transform intensities or not}

\item{...}{Passed between methods.}

\item{somaObj}{A \code{WideSomaLogicData} object to transform.}
}
\value{
ExpressionSet or MSnSet object
}
\description{
Converts objects into \code{ExpressionSet} or \code{MSnSet} instances.
}
\examples{
somaFile <- extractSampleData()
wideSomaData <- readAdat(somaFile)
as.ExpressionSet(wideSomaData)
if(requireNamespace("MSnbase"))
{
  as.MSnSet(wideSomaData)
}

unlink(somaFile)
}
\author{
Aditya Bhagwat
}

