% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment.R
\name{as.SummarizedExperiment}
\alias{as.SummarizedExperiment}
\alias{as.SummarizedExperiment.WideSomaLogicData}
\title{Convert objects into SummarizedExperiments}
\usage{
as.SummarizedExperiment(x, ...)

\method{as.SummarizedExperiment}{WideSomaLogicData}(x, ...)
}
\arguments{
\item{x}{An object to transform.  Currently only \code{WideSomaLogicData}
objects are supported.}

\item{...}{Arguments passed between methods. Currently unused.}
}
\value{
An object of classs SummarizedExperiment.
}
\description{
Convert objects into \code{SummarizedExperiments}.
}
\examples{
somaFile <- extractSampleData()
wideSomaData <- readAdat(somaFile)
as.SummarizedExperiment(wideSomaData)
unlink(somaFile)
}
