% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{chromosomalPositions}
\alias{chromosomalPositions}
\title{Chromosomal Positions by SomaLogic SeqID}
\format{A list of data frames, each with the following columns.
\describe{
 \item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{Chromsome}{Character.  Either '1' to '22' or 'X' . (Currently no 'Y' values.)}
\item{StartPosition}{Integer. Distance in base pairs from the 5' end of the gene to the start of the protein.}
\item{EndPosition}{Integer. Distance in base pairs from the 5' end of the gene to the end of the protein.}
}}
\description{
A lookup of chromosomal positions by SomaLogic Sequence ID.
}
\examples{
head(chromosomalPositions)
}

