% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAdat.R
\name{colnamesStartWithSeqId}
\alias{colnamesStartWithSeqId}
\title{Does the input contain Sequence ID column names?}
\usage{
colnamesStartWithSeqId(x)
}
\arguments{
\item{x}{A character vector.  Typically the column names of sequence data.}
}
\value{
A logical vector with the same length as \code{x}.
}
\description{
Checks if the input begins with "SeqID.".
}
\examples{
somaFile <- extractSampleData()
wideSomaData <- readAdat(somaFile)
colnamesStartWithSeqId(wideSomaData)
unlink(somaFile)
}

