% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAdat.R
\name{convertSeqIdToAptamer}
\alias{convertSeqIdToAptamer}
\title{Convert SeqIds to Aptamers}
\usage{
convertSeqIdToAptamer(seqId)
}
\arguments{
\item{seqId}{A character vector or factor of sequence IDs.}
}
\value{
A character vector of factors.
}
\description{
Converts SomaLogic sequence IDs to aptamer IDs.
}
\examples{
convertSeqIdToAptamer("2717-12_5")
}

