% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{createSampleSubmission}
\alias{createSampleSubmission}
\title{Create a SomaLogic Sample Submission Form}
\usage{
createSampleSubmission(slides, controls, comments, samples,
  sampleMatrix = c("EDTA-Plasma", "Sodium Citrate Plasma", "Serum"),
  siteId = "WCQ", studyName = "", studyId = "", runName = "Set A")
}
\arguments{
\item{slides}{A data frame of slide data, as imported by
\code{\link{readSlides}}.}

\item{controls}{A data frame of control data, as imported by
\code{\link{readControls}}.}

\item{comments}{A data frame of comment data, as imported by
\code{\link{readComments}}.}

\item{samples}{A data frame of sample data, as imported by
\code{\link{readSamples}}.}

\item{sampleMatrix}{A string giving the type of samples (plamsa or serum).}

\item{siteId}{A string giving the SomaLogic ID of your site or institution.}

\item{studyName}{A string giving your institution's name for the study.}

\item{studyId}{A string giving the SomaLogic ID of the study}

\item{runName}{A string naming the plate.}
}
\value{
A data table with the 96 rows (one plate worth, with each row
representing a sample). It contains the following columns:
\describe{
\item{SampleNumber}{One to ninety six.}
\item{SlideId}{From \code{slides.csv}.}
\item{Subarray}{Position in slide, from 1 to 8.}
\item{PlatePosition}{Position in plate, from A1 to H12}
\item{PercentDilution}{Always 40.}
\item{BarCode}{From \code{controls.csv}.}
\item{SampleNotes}{From \code{controls.csv}.}
\item{AssayNotes}{From \code{controls.csv}.}
\item{SampleId}{From \code{samples.csv}.}
\item{SampleMatrix}{\code{sampleMatrix} for samples; blank for controls.}
\item{SiteId}{\code{siteId} for samples; blank for controls.}
\item{StudyName}{\code{studyName} for samples; blank for controls.}
\item{TimePoint}{Currently blank.}
\item{SampleGroup}{Currently blank.}
\item{SampleDescription}{Currently blank.}
\item{StudyId}{\code{studyId} for samples; blank for controls.}
\item{RunName}{\code{runName} for samples; blank for controls.}
}
}
\description{
Creates a data table of contents for a SomaLogic sample submission form.
}
\seealso{
\code{\link{writeSampleSubmissionForm}} for usage examples.
}

