% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSampleData.R
\name{extractSampleData}
\alias{extractSampleData}
\title{Extract the sample datasets}
\usage{
extractSampleData(somascanMenuVersion = c("1.3k", "1.1k"),
  intermediateDir = tempdir())
}
\arguments{
\item{somascanMenuVersion}{String describing which data file to extract.
Either "1.3k" or "1.1k".}

\item{intermediateDir}{A path to a directory to extract the ADATY file into.}
}
\value{
A path to the extracted ADAT file.
}
\description{
Extracts the sample datasets from the ZIP files in the \code{extdata}
directory.
}
\examples{
extractSampleData() ##1.3k dataset by default
extractSampleData("1.1k")
}

