% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsemblIds.R
\name{getEnsemblIds}
\alias{getEnsemblIds}
\title{Get Ensembl IDs by SeqID}
\usage{
getEnsemblIds(seqIds = NULL, simplify = FALSE)
}
\arguments{
\item{seqIds}{A character vector of SomaLogic Sequence IDs, or \code{NULL} to use
all 1129 Sequence IDs.}

\item{simplify}{Logical.  Should the output be collapsed into a single
data.frame?}
}
\value{
A list of character vectors.  The names of the list are the input
Sequence IDs, and the character vector associated with that element contains the
Ensembl IDs.
}
\description{
Gets the Ensembl IDs associated with SomaLogic Sequence IDs.
}
\examples{
# Each SeqID may have one, many, or zero associated Ensembl IDs
getEnsemblIds(c("2278-61_4", "4703-87_2", "4916-2_1"))

# Get everything in the 1129 panel.
\dontrun{
getEnsemblIds()
}
}

