% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsemblIds.R
\name{getGoMolecularFunctions}
\alias{getGoBiologicalProcesses}
\alias{getGoCellularComponents}
\alias{getGoMolecularFunctions}
\title{GO definitions by SeqID}
\usage{
getGoMolecularFunctions(seqIds = NULL, simplify = FALSE)

getGoBiologicalProcesses(seqIds = NULL, simplify = FALSE)

getGoCellularComponents(seqIds = NULL, simplify = FALSE)
}
\arguments{
\item{seqIds}{A character vector of SomaLogic Sequence IDs, or \code{NULL} to
use all 1129 Sequence IDs.}

\item{simplify}{Logical.  Should the output be collapsed into a single
data.frame?}
}
\value{
A list of data frames.  The names of the list are the input
SeqIds, and the data frame associated with that element contains:
\describe{
\item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{GoId}{Character.  GO ID for property of the target protein.}
\item{GoName}{Character.  Name corresponding to the GO ID.}
\item{GoDefinition}{Character.  Description corresponding to the GO ID.}
}
}
\description{
Gets the GO definitions associated with SomaLogic Sequence IDs.  There are
three datasets, one for each of these domains: molecular function, biological
process, and cellular compartment.
}
\examples{
# Each SeqID may have one, many, or zero associated GO descriptions
getGoMolecularFunctions(c("2278-61_4", "3505-6_2", "4916-2_1"))
getGoBiologicalProcesses(c("2278-61_4", "3505-6_2", "4916-2_1"))
getGoCellularComponents(c("2278-61_4", "3505-6_2", "4916-2_1"))

# Get everything in the 1310 and 1129 panels.
\dontrun{
getGoMolecularFunctions()
getGoBiologicalProcesses()
getGoCellularComponents()
}
}

