% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{getIntensities}
\alias{as.matrix.WideSomaLogicData}
\alias{getIntensities}
\alias{getSampleData}
\title{Get the intensities or sample data}
\usage{
getIntensities(x, ...)

\method{as.matrix}{WideSomaLogicData}(x, ...)

getSampleData(x, ...)
}
\arguments{
\item{x}{An object of class \code{WideSomaLogicData} or
\code{LongSomaLogicData}.}

\item{...}{Variables passed from other methods. Currently
ignored.}
}
\value{
A numeric matrix of intensities for each protein. Row names are taken
from the \code{SampleId} of the input.  Column names are the protein
sequence IDs.
}
\description{
Gets the intensities or sample data from an object of class
\code{WideSomaLogicData} or \code{LongSomaLogicData}.
}
\examples{
\donttest{
unzip(
  system.file("extdata", "soma_atkin_diabetes.zip", package = "pdapmain"),
  exdir = tempdir()
)
soma_file <- file.path(tempdir(), "soma_atkin_diabetes.adat")
wide_soma_data <- readSomaLogic(soma_file)
unlink(soma_file)
long_soma_data <- melt(wide_soma_data)

getIntensities(wide_soma_data)         # A matrix
getIntensities(long_soma_data)         # A data.table

getSampleData(wide_soma_data)          # A data.table
getSampleData(long_soma_data)          # A data.table
}
}

