% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEnsemblIds.R
\name{getKeggDefinitions}
\alias{getKeggDefinitions}
\alias{getKeggModules}
\alias{getKeggPathways}
\title{KEGG definitions, modules, and pathways by SeqID}
\usage{
getKeggDefinitions(seqIds = NULL, simplify = FALSE)

getKeggModules(seqIds = NULL, simplify = FALSE)

getKeggPathways(seqIds = NULL, simplify = FALSE)
}
\arguments{
\item{seqIds}{A character vector of SomaLogic Sequence IDs, or \code{NULL} to
use all Sequence IDs.}

\item{simplify}{Logical.  Should the output be collapsed into a single
data.frame?}
}
\value{
A list of data frames.  The names of the list are the input
SeqIds, and the data frame associated with that element contains:
\describe{
\item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{KeggId}{Character.  KEGG ID for the gene that produces the
target protein.}
\item{KeggDefinition}{Character.  Description corresponding to the KEGG ID.}
\item{KeggCytogenicLocation}{Character.  KEGG determination of the gene's
locus.}
}
}
\description{
Gets the KEGG definitions, modules, and pathways associated with SomaLogic
Sequence IDs.
}
\examples{
# Each SeqID may have one, many, or zero associated KEGG descriptions
getKeggDefinitions(c("2278-61_4", "3505-6_2", "4916-2_1"))
getKeggModules(c("2278-61_4", "3505-6_2", "4916-2_1"))
getKeggPathways(c("2278-61_4", "3505-6_2", "4916-2_1"))

# Get everything in the 1310 and 1129 panels.
\dontrun{
getKeggDefinitions()
getKeggModules()
getKeggPathways()
}
}

