% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma2eset.R
\name{getSampleIdVar}
\alias{getSampleIdVar}
\title{Get name of sample id variable}
\usage{
getSampleIdVar(somaEset)
}
\arguments{
\item{somaEset}{eset with soma data}
}
\value{
character
}
\description{
Get name of sample id variable
}
