% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEnsemblIds.R
\name{getUniProtKeywords}
\alias{getUniProtKeywords}
\title{Get UniProt Keywords by SeqID}
\usage{
getUniProtKeywords(seqIds = NULL, simplify = FALSE)
}
\arguments{
\item{seqIds}{A character vector of SomaLogic Sequence IDs, or \code{NULL} to
use all 1129 Sequence IDs.}

\item{simplify}{Logical.  Should the output be collapsed into a single
data.frame?}
}
\value{
A list of data frames.  The names of the list are the input
SeqIds, and the data frame associated with that element contains:
\describe{
\item{UniProtId}{Character. The UniProt ID that the Keyword is associated
with.}
\item{Keyword}{Character. A UniProt Keyword associated with the SeqID and
UniProt ID.}
}
}
\description{
Gets the UniProt Keywords associated with SomaLogic Sequence IDs.
}
\examples{
# Each SeqID may have one, many, or zero associated Ensembl IDs
getUniProtKeywords(c("2278-61_4", "4703-87_2", "4916-2_1"))

# Get everything in the 1310 and 1129 panels.
\dontrun{
getUniProtKeywords()
}
}

