% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{goMolecularFunction}
\alias{goMolecularFunction}
\alias{goBiologicalProcess}
\alias{goCellularComponent}
\title{GO definitions by SomaLogic AptamerId}
\format{A list of data frames, each with the following columns.
\describe{
\item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{GoId}{Character.  GO ID for property of the target protein.}
\item{GoName}{Character.  Name corresponding to the GO ID.}
\item{GoDefinition}{Character.  Description corresponding to the GO ID.}
}}
\description{
A lookup of GO (Gene Ontology) definitions, for molecular function,
biological process, and cellular component by SomaLogic Aptamer ID.
}
\examples{
head(goMolecularFunction)
head(goBiologicalProcess)
head(goCellularComponent)
}
\references{
More information on GO can be found at:
\url{http://geneontology.org/}
}
