% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAdat.R
\name{isSeqId}
\alias{isSeqId}
\title{Does the input contain Sequence ID column names?}
\usage{
isSeqId(x)
}
\arguments{
\item{x}{A character vector.  Typically the column names of sequence data.}
}
\value{
A logical vector with the same length as \code{x}.
}
\description{
Checks if the input begins with "SeqID.".
}

