% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{keggDefinitions}
\alias{keggDefinitions}
\alias{keggModules}
\alias{keggPathways}
\alias{keggDefinitions1129}
\alias{keggModules1129}
\alias{keggPathways1129}
\title{KEGG definitions, modules, and pathways by SomaLogic AptamerId}
\format{A list of data frames, each with the following columns.
\describe{
\item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{KeggId}{Character.  KEGG ID for the gene that produces the
target protein.}
\item{KeggDefinition}{Character.  Description corresponding to the KEGG ID.}
\item{KeggCytogenicLocation}{Character.  KEGG determination of the gene's
locus.}
}}
\description{
A lookup of KEGG (Kyoto Encyclopedia of Genes and Genomes) definitions,
modules, and pathways by SomaLogic Aptamer ID.
}
\examples{
head(keggDefinitions)
head(keggModules)
head(keggPathways)
}
\references{
More information on KEGG can be found at:
\url{http://www.kegg.jp/}
}
