% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{melt.WideSomaLogicData}
\alias{melt.WideSomaLogicData}
\title{Melt a WideSomaLogicData object}
\usage{
\method{melt}{WideSomaLogicData}(data, ..., na.rm = FALSE,
  value.name = "Intensity")
}
\arguments{
\item{data}{An object of class \code{WideSomaLogicData}.}

\item{...}{Currently unused.}

\item{na.rm}{TODO}

\item{value.name}{TODO}
}
\value{
An object of class \code{LongSomaLogicData} that inherits from
\code{data.frame}.
This function melts the sample data contained in a \code{WideSomaLogicData}
object so the sequence IDs are contained in a single column \code{SeqID},
with the corresponding intensities in a single column named \code{Intensity}.
the \code{SequenceData} attribute of the input is then merged into this.
the \code{Metadata} and \code{Checksum} attributes are preserved.
}
\description{
Convert a \code{WideSomaLogicData} object from wide format to long format.
}

