% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{orderPlatePosition}
\alias{orderPlatePosition}
\title{Order Plate position}
\usage{
orderPlatePosition(pp)
}
\arguments{
\item{pp}{character vector with plate positions}
}
\value{
ordered vector
}
\description{
Order Plate position
}

