% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSomaLogic.R
\name{readAdat}
\alias{readAdat}
\alias{readSomaLogic}
\title{Read a SomaLogic data file}
\usage{
readAdat(file, keepOnlyPasses = TRUE, dateFormat = "\%Y-\%m-\%d",
  verbose = getOption("verbose"))

readSomaLogic(file, keepOnlyPasses = TRUE, dateFormat = "\%d/\%m/\%Y")
}
\arguments{
\item{file}{A string containing the path to the file to be read.}

\item{keepOnlyPasses}{A logical value indicating whether or not to keep
only the rows and columns where the data quality was considered to be
passable.}

\item{dateFormat}{A string describing the format of the dates contained in
the file's metadata.  See \code{\link[base]{strptime}} for how to specify
these.}
}
\value{
An object of class \code{WideSomaLogicData}, which inherits from
\code{data.table}.
The return value consists of a data frame where each row represents a
sample.  Initial columns contain sample metadata and later columns contain
intensities of proteins.  The specific metadata columns are not fixed, but
the most useful ones described below should always be present.
\describe{
\item{SampleUniqueID}{A unique identifier for the sample.}
\item{Subject_ID}{A unique identifier for the person being sampled.}
\item{RowCheck}{This should have the value "PASS" if the data quality is
acceptable.}
}
Columns of proteins intensities have a name beginning \code{SeqId.}.
Return value also has three attributes.
\describe{
\item{SequenceData}{A data frame.}
\item{Metadata}{A list of experimental metadata values.}
\item{Checksum}{A SHA1 checksum to ensure file integrity.}
}
}
\description{
Reads a SomaLogic ADAT data file.
}
\examples{
\donttest{
unzip(
  system.file("extdata", "soma_atkin_diabetes.zip", package = "koraproteomics"),
  exdir = tempdir()
)
soma_file <- file.path(tempdir(), "soma_atkin_diabetes.adat")
wide_soma_data <- readSomaLogic(soma_file)
str(wide_soma_data, list.len = 35)
unlink(soma_file)
}
}
\author{
Richard Cotton
}

