% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R, R/readAdat.R, R/methods.R
\name{readat-defunct}
\alias{defunct}
\alias{getSequenceInfo}
\alias{readSomaLogic}
\alias{readat-defunct}
\alias{setSequenceInfo}
\title{Defunct objects in readat}
\usage{
readSomaLogic(...)

getSequenceInfo(...)

setSequenceInfo(...)
}
\arguments{
\item{...}{Arguments to functions you shouldn't use.}
}
\value{
Possibly nonsense; don't use these functions.
}
\description{
Functions that are no longer used.
}

