% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{readComments}
\alias{readComments}
\title{Read SomaLogic Sample Submission Comments File}
\usage{
readComments(file = "comments.csv")
}
\arguments{
\item{file}{A string denoting the path to an input CSV file.  See Input file
specification section.}
}
\value{
A \code{data.table} with 96 rows and 2 columns.
\describe{
\item{PlatePosition}{A letter followed by a number, constructed from the
Subarray ("A" for 1, "B" for 2, etc.) and the slide number from 1 to 12.}
\item{SampleNotes}{Either "red", "yellow", "turbid", "red, turbid", or
"yellow, turbid", or no notes.}
\item{AssayNotes}{Free text describing assay notes, particularly problems.}
}
}
\description{
Read SomaLogic Sample Submission Comments File
}
\section{Input file specification}{

A CSV file without a header line containing up to 96 rows and three
columns as follows.
\enumerate{
\item{Plate positions. Not all positions need to be included, and they don't
need to be in order.}
\item{Sample notes. Either "red", "yellow", "turbid", "red, turbid", or
"yellow, turbid", or no notes.}
\item{Assay notes. Free text.}
}
}

\examples{
# See ?writeSampleSubmissionForm for a more complete example
withr::with_dir(
  system.file("extdata", package = "readat"),
  {
    (comments <- readComments())
  }
)
}
\seealso{
\code{\link{readSlides}}, \code{\link{readControls}}, and
\code{\link{readSamples}} for reading other submission forms and
\code{\link{writeSampleSubmissionForm}} for usage examples.
}
