% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{readControls}
\alias{readControls}
\title{Read SomaLogic Sample Submission Controls File}
\usage{
readControls(file = "controls.csv")
}
\arguments{
\item{file}{A string denoting the path to an input CSV file.  See Input file
specification section.}
}
\value{
A \code{data.table} with 96 rows and 2 columns.
\describe{
\item{PlatePosition}{A letter followed by a number, constructed from the
Subarray ("A" for 1, "B" for 2, etc.) and the slide number from 1 to 12.}
\item{BarCode}{Sample barcode for QC, Calibrator, and Buffer samples, in the
form "I" followed by 6 digits.}
}
}
\description{
Read SomaLogic Sample Submission Controls File
}
\section{Input file specification}{

A CSV file without a header line containing up to 96 rows and two
columns as follows.
\enumerate{
\item{Plate positions from A1, A2, through to H12.}
\item{Barcodes in the form "I" followed by 6 digits.}
}
}

\examples{
# See ?writeSampleSubmissionForm for a more complete example
withr::with_dir(
  system.file("extdata", package = "readat"),
  {
    (controls <- readControls())
  }
)
}
\seealso{
\code{\link{readSlides}}, \code{\link{readComments}}, and
\code{\link{readSamples}} for reading other submission forms and
\code{\link{writeSampleSubmissionForm}} for usage examples.
}
