% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSomaLogic.R
\name{readFirstChar}
\alias{readFirstChar}
\title{Read the first character of each line}
\usage{
readFirstChar(file)
}
\arguments{
\item{file}{A string or readable connection.}
}
\value{
A character vector of single characters.
}
\description{
Reads the first character of each line of a text file.
}
\references{
Richard Scriven's response to
\url{http://stackoverflow.com/q/27747426/134830}
}

