% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{readSamples}
\alias{readSamples}
\title{Read SomaLogic Sample Submission Samples File}
\usage{
readSamples(file = "samples.csv")
}
\arguments{
\item{file}{A string denoting the path to an input CSV file.  See Input file
specification section.}
}
\value{
A \code{data.table} with 96 rows and 2 columns.
\describe{
\item{PlatePosition}{A letter followed by a number, constructed from the
Subarray ("A" for 1, "B" for 2, etc.) and the slide number from 1 to 12.}
\item{SampleId}{9 digits.}
}
}
\description{
Read SomaLogic Sample Submission Samples File
}
\section{Input file specification}{

A CSV file without a header line containing up to 96 rows and at least two
columns as follows.
\enumerate{
\item{Plate positions. Not all positions need to be included, and they don't
need to be in order.}
\item{Samples IDs in the form of 9 digits, or "NO READ".}
}
Additional columns are ignored.
}

\examples{
# See ?writeSampleSubmissionForm for a more complete example
withr::with_dir(
  system.file("extdata", package = "readat"),
  {
    (samples <- readSamples())
  }
)
}
\seealso{
\code{\link{readSlides}}, \code{\link{readComments}}, and
\code{\link{readControls}} for reading other submission forms and
\code{\link{writeSampleSubmissionForm}} for usage examples.
}
