% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{readSlides}
\alias{readComments}
\alias{readControls}
\alias{readSamples}
\alias{readSlides}
\title{Read SomaLogic Sample Submission Input files}
\usage{
readSlides(file = "slides.csv")

readControls(file = "controls.csv")

readComments(file = "comments.csv")

readSamples(file = "samples.csv")
}
\arguments{
\item{file}{A string denoting the path to an input CSV file.}
}
\value{
A \code{data.table}.
}
\description{
Read SomaLogic Sample Submission Input files
}
\seealso{
\code{\link{writeSampleSubmissionForm}} for usage examples.
}

