% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma2eset.R
\name{soma2eset}
\alias{soma2eset}
\title{Convert WideSomaLogicData into ExpressionSet}
\usage{
soma2eset(somaObj, log2Transform = TRUE)
}
\arguments{
\item{somaObj}{WideSomaLogicData object}

\item{log2Transform}{whether to log2 transform intensities or not}
}
\value{
ExpressionSet object
}
\description{
Convert WideSomaLogicData into ExpressionSet
}
\author{
Aditya Bhagwat
}

