% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{uniprotKeywords}
\alias{uniprotKeywords}
\title{UniProt Keywords by SomaLogic SeqID}
\format{A list of data frames, each with the following columns.
\describe{
\item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{Keyword}{Character.  A UniProt keyword.}
}}
\description{
A lookup of UniProt keywords by SomaLogic Sequence ID.
}
\examples{
head(uniprotKeywords)
}
\references{
More information on UniProt keywords can be found at:
\url{http://www.uniprot.org/help/keywords}
}

