% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAdat.R
\name{WideSomaLogicData}
\alias{WideSomaLogicData}
\title{Create a WideSomaLogicData object}
\usage{
WideSomaLogicData(sampleAndIntensityData, sequenceData, metadata,
  checksum = paste0(rep.int("f", 40), collapse = ""))
}
\arguments{
\item{sampleAndIntensityData}{A data.table of sample and intensity data.}

\item{sequenceData}{A data.table of sequence data.}

\item{metadata}{A list of metadata.}

\item{checksum}{A string containing a SHA1 checksum.}
}
\value{
An object of class \code{WideSomaLogicData}.
}
\description{
Creates and object of class \code{WideSomaLogicData}.
}
\examples{
wsld <- WideSomaLogicData(
  data.frame(SampleStuff = letters, IntensityStuff = rnorm(26)),
  data.frame(SequenceStuff = LETTERS),
  list(MetadataStuff = Sys.Date())
)
str(wsld)
}

