% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{writeSampleSubmissionForm}
\alias{writeSampleSubmissionForm}
\title{Write a SomaLogic sample submission form}
\usage{
writeSampleSubmissionForm(submission, outdir = ".")
}
\arguments{
\item{submission}{A \code{data.frame} created by
\code{\link{createSampleSubmission}}.}

\item{outdir}{A string denoting the path to the directory where the output
file should be written.}
}
\value{
The name of the XLSX file is invisibly returned, but the function is
mostly called for the side effect of writing this file.
}
\description{
Writes a SomaLogic sample submission form to Excel XLSX file.
}
\examples{
# Import the input files
withr::with_dir(
  system.file("extdata", package = "readat"),
  {
    slides <- readSlides()
    controls <- readControls()
    comments <- readComments()
    samples <- readSamples()
  }
)

# Create the sample submission form and write to Excel spreadsheet
submission <- createSampleSubmission(
  slides, controls, comments, samples,
  studyName = "Taheri01", studyId = "WCQ-16-002"
)
writeSampleSubmissionForm(submission, tempdir())
}
\seealso{
\code{\link{readSlides}} and \code{\link{createSampleSubmission}}
}

