% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{aptamers}
\alias{aptamers}
\title{Sequence data for the Somalogic SOMAscan assay}
\format{A data frame with the following columns.
\describe{
\item{AptamerId}{Character, primary key.  The identifier of the aptamer sequence
identified using SELEX.  It consists of an SomaLogic SeqId truncated at the
underscore, to remove its sequence version number.}
\item{SomaId}{Character. The SomaLogic identifier for the protein target.
For the 1129 and 1310 assays, there is a one-to-one correspondence between
SeqId and SomaId, but in theory there is a many-to-one correspondence.}
\item{Target}{Character. The name of the protein target, from the protein
standard supplier, sometimes with additional annotation by SomaLogic.}
\item{TargetFullName}{Character. The name of the protein target, from UniProt.}
\item{UniProt}{List of character vectors. UniProt IDs for the protein target.}
\item{EntrezGeneID}{List of character vectors. Entrez Gene IDs for the gene
associated with protein target.}
\item{EntrezGeneSymbol}{List of character vectors. Entrez Gene symbols for
the gene associated with protein target.}
\item{Organism}{Either "Human" or the name of a virus.}
\item{Units}{Should always be RFU, short for Relative Fluorescence Units.}
\item{Type}{The protein standard source for the original SELEX performed for
the sequence. "Protein" refers to human protein; a few sequences used rat
protein standard sources in the SELEX.}
\item{PlasmaDilution}{What dilution factor is used in plasma?}
\item{SerumDilution}{What dilution factor is used in serum?}
\item{IsIn1129Panel}{Is the aptamer in the 1129 panel of the SOMAscan assay?}
\item{IsIn1310Panel}{Is the aptamer in the 1310 panel of the SOMAscan assay?}
}}
\description{
Sequence data for the SOMAmers in the 1129 and 1310 panels of the Somalogic
SOMAscan assay.
}
\examples{
head(aptamers)
}
\references{
The SOMAmers in the SomaLogic SOMAscan 1310 assay can be found
in this PDF:
\url{http://www.somalogic.com/somalogic/media/Assets/PDFs/SSM-045-REV-1-SOMAscan-Assay-1-3k-Content.pdf}
Those from the 1129 assay can be found here:
\url{http://www.somalogic.com/somalogic/media/Assets/PDFs/SSM-011-Rev-11-SOMAscan-Assay-\%28V1-1k\%29-Content.pdf}
}

