% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ensemblIds}
\alias{ensemblIds}
\title{Ensembl IDs by SomaLogic AptamerId}
\format{A list of character vectors.  The names of the list are SomaLogic Seq IDs, and the character vectors contain Ensembl IDs for each Seq ID.}
\description{
A lookup of Ensembl IDs by SomaLogic Aptamer ID.
}
\examples{
head(ensemblIds)
}
\references{
More information on Ensembl IDs can be found at:
\url{http://www.ensembl.org/index.html}
}
