% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getIntensities}
\alias{as.matrix.WideSomaLogicData}
\alias{getIntensities}
\alias{getIntensities.LongSomaLogicData}
\alias{getIntensities.WideSomaLogicData}
\alias{getSampleData}
\alias{getSampleData.LongSomaLogicData}
\alias{getSampleData.WideSomaLogicData}
\title{Get the intensities or sample data}
\usage{
getIntensities(x, ...)

\method{getIntensities}{WideSomaLogicData}(x, rowsContain = c("samples",
  "sequences"), reorder = FALSE, ...)

\method{getIntensities}{LongSomaLogicData}(x, ...)

\method{as.matrix}{WideSomaLogicData}(x, ...)

getSampleData(x, ...)

\method{getSampleData}{WideSomaLogicData}(x, ...)

\method{getSampleData}{LongSomaLogicData}(x, ...)
}
\arguments{
\item{x}{An object of class \code{WideSomaLogicData} or
\code{LongSomaLogicData}.}

\item{...}{Variables passed to and from from other methods.}

\item{rowsContain}{Either samples or sequences.}
}
\value{
A numeric matrix of intensities for each protein. Row names are taken
from the \code{ExtIdentifer} of the input.  Column names are the protein
sequence IDs.
}
\description{
Gets the intensities or sample data from an object of class
\code{WideSomaLogicData} or \code{LongSomaLogicData}.
}
\examples{
\donttest{
soma_file <- extractSampleData()
wide_soma_data <- readAdat(soma_file)

intWideSamp <- getIntensities(wide_soma_data) # A matrix
View(intWideSamp, "Wide intensities, samples per row")

intWideSeq <- getIntensities(                 # The transpose
  wide_soma_data,
  rowsContain = "sequences"
)
View(intWideSeq, "Wide intensities, seqs per row")

sampWide <- getSampleData(wide_soma_data)     # A data.table
View(sampWide, "Wide sample data")

if(requireNamespace("reshape2"))
{
  long_soma_data <- reshape2::melt(wide_soma_data)
  intLong <- getIntensities(long_soma_data)   # A data.table
  View(intLong, "Long intensities")

  sampLong <- getSampleData(long_soma_data)   # A data.table
  View(sampLong, "Long sample data")
}
}
}

