% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-annotations.R
\name{getKeggDefinitions}
\alias{getKeggDefinitions}
\alias{getKeggModules}
\alias{getKeggPathways}
\title{KEGG definitions, modules, and pathways by AptamerId}
\usage{
getKeggDefinitions(aptamerIds = NULL, simplify = FALSE)

getKeggModules(aptamerIds = NULL, simplify = FALSE)

getKeggPathways(aptamerIds = NULL, simplify = FALSE)
}
\arguments{
\item{aptamerIds}{A character vector of SomaLogic aptamer IDs, or \code{NULL}
to use all aptamer IDs.}

\item{simplify}{Logical.  Should the output be collapsed into a single
data.frame?}
}
\value{
A list of data frames.  The names of the list are the input
aptamerIds, and the data frame associated with that element contains:

\describe{
\item{UniProtId}{Character.  UniProt ID for the protein target.}
\item{KeggId}{Character.  KEGG ID for the gene that produces the
target protein.}
\item{KeggDefinition}{Character.  Description corresponding to the KEGG ID.}
\item{KeggCytogenicLocation}{Character.  KEGG determination of the gene's
locus.}
}
}
\description{
Gets the KEGG definitions, modules, and pathways associated with SomaLogic
aptamer IDs.
}
\examples{
# Each AptamerId may have one, many, or zero associated KEGG descriptions
getKeggDefinitions(c("2278-61", "3505-6", "4916-2"))
getKeggModules(c("2278-61", "3505-6", "4916-2"))
getKeggPathways(c("2278-61", "3505-6", "4916-2"))

# Get everything in the 1310 and 1129 panels.
\dontrun{
getKeggDefinitions()
getKeggModules()
getKeggPathways()
}
}
