% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-annotations.R
\name{getPfam}
\alias{getPfam}
\title{Get PFAM IDs by AptamerId}
\usage{
getPfam(aptamerIds = NULL, simplify = FALSE)
}
\arguments{
\item{aptamerIds}{A character vector of SomaLogic aptamer IDs, or \code{NULL}
to use all aptamer IDs.}

\item{simplify}{Logical.  Should the output be collapsed into a single
data.frame?}
}
\value{
A list of data frames.  The names of the list are the input
SeqIds, and the data frame associated with that element contains:
\describe{
\item{EntrezGeneId}{Character.  EntrezGene IDs for the gene that produces
the target protein.}
\item{PfamId}{Character.  PFAM ID for a property of the target protein.}
\item{PfamDescription}{Character.  Description of a PFAM protein property.}
}
}
\description{
Gets the PFAM Ids and descriptions associated with SomaLogic aptamer IDs.
}
\examples{
# Each AptamerId may have one, many, or zero associated PFAM descriptions
getPfam(c("2278-61", "4703-87", "4916-2"))

# Get everything in the 1310 and 1129 panels.
\dontrun{
getPfam()
}
}
