% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequencesWithLargestBetweenGroupVariation.R
\name{getSequencesWithLargestBetweenGroupVariation}
\alias{getSequencesWithLargestBetweenGroupVariation}
\title{Get sequences with the largest between group variation}
\usage{
getSequencesWithLargestBetweenGroupVariation(x, n = 10,
  group = ~SampleGroup, ...)
}
\arguments{
\item{x}{An object of class LongSomaLogicData.}

\item{n}{An integer of the number of sequences to return, passed to
\code{\link[dplyr]{top_n}}.}

\item{group}{A formula, string or quoted column name of the column that
defines the groups to split by.}

\item{...}{Passed to and from methods, but currently unused.}
}
\value{
A data table with \code{n} rows and the following columns.
\describe{
  \item{SeqId}{SomaLogic sequence identifier.}
  \item{VariationBetweenGroups}{The largest mean log intensity within a
  \code{group} divided by the smallest mean log intensity within a
  \code{group}.}
}
}
\description{
Get the sequences with the largest between group variation in mean log
intensity.
}
