% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{readat-defunct}
\alias{defunct}
\alias{readat-defunct}
\title{Defunct objects in readat}
\arguments{
\item{...}{Arguments to functions you shouldn't use.}
}
\value{
Possibly nonsense; don't use these functions.
}
\description{
Functions that are no longer used.
}

