% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sample-submision-input-files.R
\name{readSlides}
\alias{readSlides}
\title{Read SomaLogic Sample Submission Slides File}
\usage{
readSlides(file = "slides.csv")
}
\arguments{
\item{file}{A string denoting the path to an input CSV file.  See Input file
specification section.}
}
\value{
A \code{data.table} with 96 rows and 5 columns.
\describe{
\item{SampleNumber}{The integers 1 to 96.}
\item{SlideId}{The slide IDs from the input file.}
\item{Subarray}{Integers from 1 to 8 denoting the sample position for the
slide.}
\item{PlatePosition}{A letter followed by a number, constructed from the
Subarray ("A" for 1, "B" for 2, etc.) and the slide number from 1 to 12.}
\item{PercentDilution}{Always 40.}
}
}
\description{
Read SomaLogic Sample Submission Slides File
}
\note{
For partially filled plates (where there are less than 96 samples),
there isn't enough information in the file to programmatically determine
which slide IDs correspond to which sample IDs.  In this case please manually
create the slides dataset.
}
\section{Input file specification}{

A CSV file without a header line containing up to twelve rows and one
column as follows.
\enumerate{
\item{Slide IDs, each 12 digits long.}
}
}

\examples{
# See ?writeSampleSubmissionForm for a more complete example
withr::with_dir(
  system.file("extdata", package = "readat"),
  {
    (slides <- readSlides())
  }
)
}
\seealso{
\code{\link{readControls}}, \code{\link{readComments}}, and
\code{\link{readSamples}} for reading other submission forms and
\code{\link{writeSampleSubmissionForm}} for usage examples.
}
