% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSomaLogic.R
\name{WideSomaLogicDataAttributes}
\alias{WideSomaLogicDataAttributes}
\alias{getChecksum}
\alias{getMetadata}
\alias{getSequenceInfo}
\alias{setChecksum}
\alias{setMetadata}
\alias{setSequenceInfo}
\title{Get WideSomaLogicData attributes}
\usage{
getSequenceInfo(x)

getMetadata(x)

getChecksum(x)

setSequenceInfo(x, value)

setMetadata(x, value)

setChecksum(x, value)
}
\arguments{
\item{x}{An object of class \code{WideSomaLogicData}.}

\item{value}{Value to set the attribute to.}
}
\value{
An attribute of the input.
For inputs that are not \code{WideSomaLogicData} objects, the return value
may be \code{NULL}.
}
\description{
Shortcut functions to get attributes of \code{WideSomaLogicData} objects.
}
\seealso{
\code{\link{readSomaLogic}}
}

