% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getIntensities}
\alias{WideSomaLogicDataAttributes}
\alias{as.matrix.WideSomaLogicData}
\alias{getChecksum}
\alias{getIntensities}
\alias{getIntensities.LongSomaLogicData}
\alias{getIntensities.WideSomaLogicData}
\alias{getMetadata}
\alias{getSampleData}
\alias{getSampleData.LongSomaLogicData}
\alias{getSampleData.WideSomaLogicData}
\alias{getSequenceData}
\alias{setChecksum}
\alias{setIntensities}
\alias{setMetadata}
\alias{setSampleData}
\alias{setSequenceData}
\title{Get WideSomaLogicData attributes}
\usage{
getIntensities(x, ...)

\method{getIntensities}{WideSomaLogicData}(x, rowsContain = c("samples",
  "sequences"), reorder = FALSE, ...)

\method{getIntensities}{LongSomaLogicData}(x, ...)

\method{as.matrix}{WideSomaLogicData}(x, ...)

getSampleData(x, ...)

\method{getSampleData}{WideSomaLogicData}(x, ...)

\method{getSampleData}{LongSomaLogicData}(x, ...)

getSequenceData(x)

getMetadata(x)

getChecksum(x)

setSequenceData(x, value)

setMetadata(x, value)

setChecksum(x, value)

setSampleData(x, value)

setIntensities(x, value, prependSeqIdToColNames = NA)
}
\arguments{
\item{x}{An object of class \code{WideSomaLogicData} or
\code{LongSomaLogicData}.}

\item{...}{Variables passed to and from from other methods.}

\item{rowsContain}{Either samples or sequences.}

\item{reorder}{If \code{TRUE}, rows are reordered by \code{ExtIdentifier} and
columns are reordered by \code{SeqId}, alphabetically.}

\item{value}{Value to set the attribute to.}

\item{prependSeqIdToColNames}{Logical.  Should "SeqId." be prepended to the
column names of the intensities? If \code{NA}, auto-guess whether they
should be.}
}
\value{
A numeric matrix of intensities for each protein. Row names are taken
from the \code{ExtIdentifer} of the input.  Column names are the protein
sequence IDs.
}
\description{
Accessors and mutators (getters and setters) for objects of class
\code{WideSomaLogicData} or \code{LongSomaLogicData}.
}
\examples{
\donttest{
# Get the sample dataset
soma_file <- extractSampleData()
wide_soma_data <- readAdat(soma_file)

# Access its components
checksum <- getChecksum(wide_soma_data)
metadata <- getMetadata(wide_soma_data)
sequenceData <- getSequenceData(wide_soma_data)
sampleData <- getSampleData(wide_soma_data)

# Intensities of a WideSomaLogicData object are a matrix
intWideSamp <- getIntensities(wide_soma_data)
\dontrun{
  View(intWideSamp, "Wide intensities, samples per row")
}
intWideSeq <- getIntensities(                 # The transpose
  wide_soma_data,
  rowsContain = "sequences"
)
\dontrun{
  View(intWideSeq, "Wide intensities, seqs per row")
}

# Sample data is always a data table
sampWide <- getSampleData(wide_soma_data)
\dontrun{
  View(sampWide, "Wide sample data")
}

if(requireNamespace("reshape2"))
{
  # For LongSomaLogicData objects, the intensities are returned
  # as a data.table
  long_soma_data <- reshape2::melt(wide_soma_data)
  intLong <- getIntensities(long_soma_data)
  \dontrun{
    View(intLong, "Long intensities")
  }

  # Sample data has a different shape now
  sampLong <- getSampleData(long_soma_data)
  \dontrun{
    View(sampLong, "Long sample data")
  }
}
}
}
\seealso{
\code{\link{readAdat}}
}

