% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getIntensities}
\alias{getIntensities}
\alias{getIntensities.WideSomaLogicData}
\alias{getIntensities.LongSomaLogicData}
\alias{as.matrix.WideSomaLogicData}
\alias{getSampleData}
\alias{getSampleData.WideSomaLogicData}
\alias{getSampleData.LongSomaLogicData}
\alias{WideSomaLogicDataAttributes}
\alias{getSequenceData}
\alias{getMetadata}
\alias{getChecksum}
\alias{setSequenceData}
\alias{setMetadata}
\alias{setChecksum}
\alias{setSampleData}
\alias{setIntensities}
\title{Get WideSomaLogicData attributes}
\usage{
getIntensities(x, ...)

\method{getIntensities}{WideSomaLogicData}(x, rowsContain = c("samples",
  "sequences"), reorder = FALSE, ...)

\method{getIntensities}{LongSomaLogicData}(x, ...)

\method{as.matrix}{WideSomaLogicData}(x, ...)

getSampleData(x, ...)

\method{getSampleData}{WideSomaLogicData}(x, ...)

\method{getSampleData}{LongSomaLogicData}(x, ...)

getSequenceData(x)

getMetadata(x)

getChecksum(x)

setSequenceData(x, value)

setMetadata(x, value)

setChecksum(x, value)

setSampleData(x, value)

setIntensities(x, value, prependSeqIdToColNames = NA)
}
\arguments{
\item{x}{An object of class \code{WideSomaLogicData} or
\code{LongSomaLogicData}.}

\item{...}{Variables passed to and from from other methods.}

\item{rowsContain}{Either samples or sequences.}

\item{reorder}{If \code{TRUE}, rows are reordered by \code{ExtIdentifier} and
columns are reordered by \code{SeqId}, alphabetically.}

\item{value}{Value to set the attribute to.}

\item{prependSeqIdToColNames}{Logical.  Should "SeqId." be prepended to the
column names of the intensities? If \code{NA}, auto-guess whether they
should be.}
}
\value{
\code{getIntensities} returns a numeric matrix of intensities for
each protein. Row names are taken from the \code{ExtIdentifier} of the input.
Column names are the protein sequence IDs.  (If you set
\code{rowsContain = "sequences"}, then rows and columns and their names are
swapped.)

\code{getSampleData} returns a data table containing the sample data.  That
is, the input without the intensities or attributes.  There is one compulsory
column:
\describe{
\item{ExternalId}{Character. A unique identifer for the sample.}
}
These columns are not compulsory for the file spec, but are always included by SomaLogic:
\describe{
\item{SampleId}{Character. The customer's original sample identifier. This is either the same as the subject identifier, or the calibrator and buffer separated by a hyphen. For example "1234" or "PPT-09".}
\item{TimePoint}{Character or numeric.  The time point identified by the customer.}
\item{SampleGroup}{Character. Cohort information provided by the customer.}
\item{SampleNotes}{Character. Lab comments about the sample condition.}
\item{AssayNotes}{Character. Lab comments about assay adverse events.}
}
Common optional columns:
\describe{
\item{PlateId}{Character. Identifier for the plate. For assay experiments,
this also functions as an experiment identifier.}
\item{SlideId}{Character. Agilent slide barcode.}
\item{Subarray}{Integer. Agilent subarray number from 1 to 8.}
\item{SampleType}{Character. Either "Sample, "QC", "Buffer" or "Calibrator".}
\item{BarCode}{Character. 1D Barcode of aliquot.}
\item{Barcode2d}{Character. 2D Barcode of aliquot.}
\item{SiteId}{Character. The laboratory that ran the experiment, applicable
if subcontracted.}
\item{SampleUniqueId}{Character.  Where multiple experiments are combined,
this column can be used to ensure a unique sample ID.}
\item{Subject_ID}{Character. Identifier for the person/animal/thing being
sampled.}
\item{HybControlNormScale}{Numeric. Hybridization control normalization scale
factor. For example, 1.23.}
\item{NormScale_40}{Numeric. Median normalization scale factor, for only the
sequences diluted to 40\% concentration. For example, 1.23.}
\item{NormScale_0_005}{Numeric. Median normalization scale factor, for only
the sequences diluted to 0.005\% concentration. For example, 1.23.}
\item{NormScale_1}{Numeric. Median normalization scale factor, for only the
sequences diluted to 1\% concentration. For example, 1.23.}
\item{PercentDilution}{Numeric. How much was the sample diluted? Either 40,
1, or 0.005.}
\item{SampleMatrix}{Character. Either "EDTA-Plasma", "Sodium Citrate Plasma"
or "Serum". Applicable if different matrices are used with different samples
in the file, otherwise use the StudyMatrix metadata value.}
\item{SampleDescription}{Character. Free text describing the sample.}
\item{TubeUniqueID}{Character. A unique identifier for every sample tube,
assigned by the customer.}
\item{RowCheck}{Character. Either "PASS" or "FAIL". Did the sample pass
quality control checks?}
}

\code{getSequenceData} returns a data table containing the sequence data.
"Sequence" can be considered synonymous with "feature" or "SOMAmer reagent".
The following columns are compulsory.
\describe{
\item{SeqId}{A unique identifer for the SOMAmer reagent.}
\item{Target}{ The unique name for the targeted proteins.}
}
The following columns are optional but should always be provided by
SomaLogic.
\describe{
\item{SomaId}{Character.  The SomaLogic identifier for the protein target.
For the 1129 and 1310 assays, there is a one-to-one correspondence between
SeqId and SomaId, but in theory there is a many-to-one correspondence.}
\item{TargetFullName}{Character.  A description of the proteins targeted by
the SOMAmer reagent, taken from UniProt.}
\item{UniProt}{Character. The UniProt identifiers for the targeted proteins.
Older versions of the file format may also contain the value 'Family',
referring to a whole family of proteins, though this behaviour is considered
deprecated.}
\item{EntrezGeneID}{Character. The Entrez Gene identifiers for the genes
corresponding to the targeted proteins.}
\item{EntrezGeneSymbol}{Character. The Entrez Gene symbols for the genes
corresponding to the targeted proteins.}
\item{Organism}{Character. What animal does the target protein refer to?}
\item{ColCheck}{Character. Either PASS or FAIL. Did the sequence pass quality
control checks?}
\item{Units}{Character.  Unit of measurement for the SOMAmer reagent
abundances. Should always be RFU.}
\item{Type}{Character. One of "Spuriomer", "Protein",
"Hybridization Control", "Non-human Protein".}
\item{Cal_*PlateId*}{Numeric.  Calibration scale factor for each plate. For
example, 1.23.}
\item{CalReference}{Numeric. Calibration reference intensity, in RFU. For
example, 543.21.}
\item{Dilution}{Numeric. Concentration of the diluted sample compared to the
neat sample, as a percentage. Either 40, 1, or 0.005.}
}

\code{getCheckSum} and \code{getMetadata} return the checksum (a string) and
metadata (a list) attributes of the input.
}
\description{
Accessors and mutators (getters and setters) for objects of class
\code{WideSomaLogicData} or \code{LongSomaLogicData}.
}
\examples{
# Get the sample dataset
soma_file <- extractSampleData()
wide_soma_data <- readAdat(soma_file)

# Access its components
checksum <- getChecksum(wide_soma_data)
metadata <- getMetadata(wide_soma_data)
sequenceData <- getSequenceData(wide_soma_data)
sampleData <- getSampleData(wide_soma_data)

# Intensities of a WideSomaLogicData object are a matrix
intWideSamp <- getIntensities(wide_soma_data)
\donttest{
# Not run due to side effects of View
  View(intWideSamp, "Wide intensities, samples per row")
}
intWideSeq <- getIntensities(                 # The transpose
  wide_soma_data,
  rowsContain = "sequences"
)
\donttest{
  View(intWideSeq, "Wide intensities, seqs per row")
}

# Sample data is always a data table
sampWide <- getSampleData(wide_soma_data)
\donttest{
  View(sampWide, "Wide sample data")
}

# For LongSomaLogicData objects, the intensities are returned
# as a data.table
long_soma_data <- reshape2::melt(wide_soma_data)
intLong <- getIntensities(long_soma_data)
\donttest{
  View(intLong, "Long intensities")
}

# Sample data has a different shape now
sampLong <- getSampleData(long_soma_data)
\donttest{
  View(sampLong, "Long sample data")
}
}
\seealso{
\code{\link{readAdat}}
}
